maint: Silence compiler warning about combining Qt key sequences.
* gui-preferences-sc.h (OCTAVE_QT_KEYCOMBINATION): When using Qt5, #include
<cstdint>, and cast inputs to macro to uint32_t.

* octave-qscintilla.h (OCTAVE_QT_KEYCOMBINATION): New macro to use
QKeyCombination class from Qt6, or with Qt5 cast inputs to uint32_t and use
logical or ('|').

* octave-qscintilla.cc (show_replace_action_tooltip): Use macro
OCTAVE_QT_KEYCOMBINATION to combine Qt modifier and Qt key.

+++ libgui/src/gui-preferences-sc.h
@@ -41,14 +41,14 @@
 // Therefore, these prefs for key sequences require a separate constant
 // definition and value method for the settings class.
 
-
 #if defined (HAVE_QKEYCOMBINATION_CLASS)
 #  include <QKeyCombination>
 #  define OCTAVE_QT_KEYCOMBINATION(mod, key) \
      QKeyCombination (mod, key).toCombined ()
 #else
+#  include <cstdint>
 #  define OCTAVE_QT_KEYCOMBINATION(mod, key) \
-     mod | key
+     static_cast<uint32_t> (mod) | static_cast<uint32_t> (key)
 #endif
 
 #if defined (Q_OS_MAC)
